requirements.txt
scipy 
numpy 
scikit-learn 
matplotlib 
pandas
pillow 
opencv-python
fire
tqdm
